initial commit; chore: initial React structure
This is the first commit for my React Neighborhood Map, final project for the Udacity FEWD Nanodegree.

With this commit, the initial structure for React components was initialized, intended as follows:
- App
- - Map
- - - MapMarker
- - - - MapWindow
- - - - - WikiApi
- - Menu
- - - MenuFilter
- - - MenuList

A temporary name was given to the app:
ReaCT
A play on words with React and CT, short for Catania, my city, which I'm gonna use to test it.