style: responsive css
The components are provided with some filler, static code to simulate the app behaviour, so that the CSS could be edited to implement responsive design.

A logo for the app and a hamburger menu icon were added as well.

- minor changes:
Two utility APIs are planned: one to fetch the locations used by the app (probably from a static JSON file), and one to get infos on each location from a third-party application (as of now, it is intended for Wikipedia).

- TODO:
1. Add markers and info windows with filler text and revise the CSS
2. Implement the APIs and the fetch-cache cycle
3. Add accessibility to the app
4. Revise HTML, CSS and JS/JSX according to Udacity's standards
5. Add documentation and comments where needed