feat: accessibility
The website has been thoroughly revised in regards to the accessibility standards, testing the functionalities with the VoiceOver software.
Navigation through the website is well described at each step, subdivided in logical groups (main presentation, query form, list of locations, info box on the map) so that the user can skip useless navigation, and links to move back and forth to specific parts of the page are provided, as visually-hidden links.

Also, some work has been done on the minor TODOs from the previous commit.

TODO:
1. Comment and document the code where needed
2. Revise the code to adhere to Udacity's standards

Final TODOs:
- Add a service worker for caching
- Compare the app with the project Rubric's requirements