fix: re-submit: meet specifications
To meet the final missing specification, I had to provide clear functionalities in case a fetch operation failed.
In order to do so, I refactored the code for the Map, letting the App component append the script and check its 'onload' and 'onerror' functions with a variable in the state.
So tha Map can render either:
- a "Loading" screen (light blue)
- an "Error" screen, with immediate focus for screen readers (red)
- render the map normally
Same kind of work with the article from Wikipedia, with a loading, error and normal state.

Additionally, content to run the app in Production mode was added and the README updated